py/bin/requirements.txt
flask
Flask-Login==0.5.0
pycryptodome
